+++ src/main/java/org/lms/controller/QuizRestController.java
@@ -1,5 +1,6 @@
 package org.lms.controller;
 
+import org.lms.AuthorizationManager;
 import org.lms.entity.Quiz;
 import org.lms.service.QuizService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -13,9 +14,13 @@ import java.util.List;
 public class QuizRestController {
     @Autowired
     private QuizService quizService;
-
+    @Autowired
+    private AuthorizationManager authorizationManager;
     @PutMapping("/")
     public ResponseEntity<String> createQuiz(@PathVariable("courseid") Long courseId, @RequestBody Quiz quiz) {
+        if (!authorizationManager.checkCourseEdit(courseId)) {
+            return ResponseEntity.status(403).body("You do not have permission to edit this course.");
+        }
         if (quizService.create(courseId,quiz)){
             return ResponseEntity.ok("All good!");
         }
@@ -24,6 +29,9 @@ public class QuizRestController {
 
     @PatchMapping("/{id}/")
     public ResponseEntity<String> updateQuiz(@PathVariable("courseid") Long courseId, @PathVariable("id") Long id, @RequestBody Quiz quiz, @PathVariable String courseid) {
+        if (!authorizationManager.checkCourseEdit(courseId)) {
+            return ResponseEntity.status(403).body("You do not have permission to edit this course.");
+        }
         if (quizService.update(courseId, id,quiz)){
             return ResponseEntity.ok("All good!");
         }
@@ -31,6 +39,9 @@ public class QuizRestController {
     }
     @GetMapping("/{id}/")
     public ResponseEntity<Quiz> getById(@PathVariable("courseid") Long courseId, @PathVariable("id") Long id) {
+        if (!authorizationManager.checkCourseView(courseId)) {
+            return ResponseEntity.status(403).body(null);
+        }
         Quiz quiz = quizService.getById(courseId,id);
         if (quiz == null) {
             return ResponseEntity.notFound().build();
@@ -38,11 +49,21 @@ public class QuizRestController {
         return ResponseEntity.ok(quiz);
     }
     @GetMapping("/")
-    public List<Quiz> getAll(@PathVariable("courseid") Long courseId) {
-        return quizService.getAll(courseId);
+    public ResponseEntity<List<Quiz>> getAll(@PathVariable("courseid") Long courseId) {
+        if (!authorizationManager.checkCourseView(courseId)) {
+            return ResponseEntity.status(403).body(null);
+        }
+        List<Quiz> quizzes = quizService.getAll(courseId);
+        if (quizzes != null && !quizzes.isEmpty()) {
+            return ResponseEntity.ok(quizzes);
+        }
+        return ResponseEntity.ok().body(quizzes);
     }
     @DeleteMapping("/{id}/")
-    public ResponseEntity<String> deleteQuiz(@PathVariable("courseid") Long courseId,@PathVariable("id") Long id) {
+    public ResponseEntity<String> delete(@PathVariable("courseid") Long courseId,@PathVariable("id") Long id) {
+        if (!authorizationManager.checkCourseEdit(courseId)) {
+            return ResponseEntity.status(403).body("You do not have permission to edit this course.");
+        }
         if(quizService.delete(courseId,id)){
             return ResponseEntity.ok("All good!");
         }
